requirements.txt
langgraph==0.3.34
google-generativeai==0.8.5
arxiv
requests
numpy
beautifulsoup4
chromadb
ipython
jupyter
matplotlib
pandas 